requirements.txt
django==2.2.5
djangorestframework==3.10.3
tqdm==4.36.1
tensorflow==1.14.0
keras==2.2.5
google-cloud-vision
django-drf-filepond
django-cors-headers
